backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
marshmallow==3.20.2
marshmallow-sqlalchemy==1.1.1
Flask-Migrate==4.0.5
Flask-Cors==3.0.10
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn
requests
Flask-SocketIO==6.4.0 